requirements.txt
PySide6>=6.7.0
PySide6-Addons>=6.7.0
markdown>=3.6
pygments>=2.17.0
watchdog>=4.0.0
toml>=0.10.2
PyYAML>=6.0
